env_setup_utils/pyproject.toml
[tool.poetry]
name = "env-setup-utils"
version = "0.1.3"
description = "Various utilities from Environment Setup project."
authors = [
    "Anonymous authors"
]
packages = [
    { include = "env_setup_utils" },
    { include = "env_setup_utils/py.typed" },
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
mistune = "^3.0.2"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.12.1"
jsonlines = "^4.0.0"
huggingface-hub = "^0.24.7"
wandb = "^0.19.5"
hydra-core = "^1.3.2"
rich = "^13.9.4"
python-dotenv = "^1.0.1"

[tool.poetry.group.analysis]
optional = true

[tool.poetry.group.analysis.dependencies]
wandb = ">=0.18.0"
plotly = "^5.18.0"
flask = "^3.0.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.0"
pytest = "^7.4.0"
ruff = "^0.3.2"
pyright = "^1.1.367"
isort = "^5.13.2"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
extend-select = ["I"]

[tool.isort]
profile = "black"
force_sort_within_sections = true
order_by_type = true

[tool.mypy]
python_version = "3.10"

[[tool.mypy.overrides]]
module = []
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"